pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-work-summary-assistant"
version = "0.1.0"
authors = [
    { name = "Wang Lei", email = "327714319@qq.com" },
]
description = "智能工作总结助手 - 基于AI的周报生成工具"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "fastapi>=0.103.1",
    "uvicorn>=0.23.2",
    "transformers>=4.33.1",
    "torch>=2.0.1",
    "pydantic>=2.3.0",
    "pydantic_settings",
    "python-dotenv>=1.0.0",
    "Django>=5.0.2",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.1",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "sentencepiece",
    "tiktoken",
    "protobuf",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "ipython",
    "django-extensions",
    "pre-commit",
]

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
